moved htmlUtil
