+++ tests/Util/Html/HtmlUtilTest.php
@@ -11,7 +11,7 @@ namespace HeimrichHannot\UtilsBundle\Tests\Util\Html;
 use HeimrichHannot\UtilsBundle\Tests\AbstractUtilsTestCase;
 use HeimrichHannot\UtilsBundle\Util\HtmlUtil\GenerateDataAttributesStringArrayHandling;
 use HeimrichHannot\UtilsBundle\Util\HtmlUtil\GenerateDataAttributesStringOptions;
-use HeimrichHannot\UtilsBundle\Util\HtmlUtil\HtmlUtil;
+use HeimrichHannot\UtilsBundle\Util\HtmlUtil;
 use PHPUnit\Framework\MockObject\MockBuilder;
 
 class HtmlUtilTest extends AbstractUtilsTestCase
